refactor: split hook managed entry helpers
