+++ ClaudeAtoll/Services/Hooks/HookInstaller.swift
@@ -11,7 +11,6 @@ import os.log
 
 // MARK: - HookInstaller
 
-// swiftlint:disable type_body_length
 /// Hook installer — MainActor (default) protects static mutable state
 /// This ensures thread-safe access to detectedRuntime across all call sites
 enum HookInstaller {
@@ -20,6 +19,12 @@ enum HookInstaller {
     /// Cached detected runtime for command generation
     /// Protected by @MainActor isolation to prevent data races
     private(set) static var detectedRuntime: PythonRuntimeDetector.PythonRuntime?
+
+    // MARK: Managed Hooks
+
+    static let hookScriptName = "claude-atoll-state.py"
+    static let legacyHookScriptName = "claude-island-state.py"
+    static let managedHookScriptNames = [hookScriptName, legacyHookScriptName]
 
     /// Install hook script and update settings.json on app launch
     /// Supports cooperative cancellation - checks Task.isCancelled at key points
@@ -131,11 +136,36 @@ enum HookInstaller {
         }
     }
 
-    // MARK: Private
+    // MARK: Managed Hook Configuration
 
-    private static let hookScriptName = "claude-atoll-state.py"
-    private static let legacyHookScriptName = "claude-island-state.py"
-    private static let managedHookScriptNames = [hookScriptName, legacyHookScriptName]
+    /// Build hook configurations for all events
+    static func buildHookConfigurations(command: String) -> [(String, [[String: Any]])] {
+        let hookEntry: [[String: Any]] = [["type": "command", "command": command]]
+        let hookEntryWithTimeout: [[String: Any]] = [["type": "command", "command": command, "timeout": 86400]]
+        let withMatcher: [[String: Any]] = [["matcher": "*", "hooks": hookEntry]]
+        let withMatcherAndTimeout: [[String: Any]] = [["matcher": "*", "hooks": hookEntryWithTimeout]]
+        let withoutMatcher: [[String: Any]] = [["hooks": hookEntry]]
+        let preCompactConfig: [[String: Any]] = [
+            ["matcher": "auto", "hooks": hookEntry],
+            ["matcher": "manual", "hooks": hookEntry],
+        ]
+
+        // TODO(anthropics/claude-code#15897): Re-add ("PreToolUse", withMatcher) once upstream
+        // fixes parallel hook updatedInput aggregation. Removed to prevent rtk interference.
+        return [
+            ("UserPromptSubmit", withoutMatcher),
+            ("PostToolUse", withMatcher),
+            ("PermissionRequest", withMatcherAndTimeout),
+            ("Notification", withMatcher),
+            ("Stop", withoutMatcher),
+            ("SubagentStop", withoutMatcher),
+            ("SessionStart", withoutMatcher),
+            ("SessionEnd", withoutMatcher),
+            ("PreCompact", preCompactConfig),
+        ]
+    }
+
+    // MARK: Private
 
     nonisolated private static let logger = Logger(subsystem: "com.engels74.ClaudeAtoll", category: "HookInstaller")
 
@@ -215,229 +245,7 @@ enum HookInstaller {
             json["hooks"] = hooks
         }
     }
-
-    /// Build hook configurations for all events
-    private static func buildHookConfigurations(command: String) -> [(String, [[String: Any]])] {
-        let hookEntry: [[String: Any]] = [["type": "command", "command": command]]
-        let hookEntryWithTimeout: [[String: Any]] = [["type": "command", "command": command, "timeout": 86400]]
-        let withMatcher: [[String: Any]] = [["matcher": "*", "hooks": hookEntry]]
-        let withMatcherAndTimeout: [[String: Any]] = [["matcher": "*", "hooks": hookEntryWithTimeout]]
-        let withoutMatcher: [[String: Any]] = [["hooks": hookEntry]]
-        let preCompactConfig: [[String: Any]] = [
-            ["matcher": "auto", "hooks": hookEntry],
-            ["matcher": "manual", "hooks": hookEntry],
-        ]
-
-        // TODO(anthropics/claude-code#15897): Re-add ("PreToolUse", withMatcher) once upstream
-        // fixes parallel hook updatedInput aggregation. Removed to prevent rtk interference.
-        return [
-            ("UserPromptSubmit", withoutMatcher),
-            ("PostToolUse", withMatcher),
-            ("PermissionRequest", withMatcherAndTimeout),
-            ("Notification", withMatcher),
-            ("Stop", withoutMatcher),
-            ("SubagentStop", withoutMatcher),
-            ("SessionStart", withoutMatcher),
-            ("SessionEnd", withoutMatcher),
-            ("PreCompact", preCompactConfig),
-        ]
-    }
-
-    /// Update existing hook entries or add new ones, deduplicating managed entries by matcher
-    private static func updateOrAddHookEntries(
-        existing: [[String: Any]]?,
-        config: [[String: Any]],
-        command: String,
-        eventName: String,
-    ) -> [[String: Any]] {
-        guard var existingEvent = existing else {
-            return config
-        }
-
-        // First, remove any legacy direct format entries (not wrapped in "hooks")
-        existingEvent.removeAll { self.isLegacyDirectEntry($0) }
-
-        // Deduplicate and update managed entries, preserving user hooks
-        let (updatedEntries, seenMatchers) = self.deduplicateClaudeAtollEntries(
-            in: existingEvent, command: command, eventName: eventName,
-        )
-        existingEvent = updatedEntries
-
-        // Add any missing configurations (matchers not already present)
-        for configEntry in config {
-            let configMatcher = (configEntry["matcher"] as? String) ?? ""
-            if !seenMatchers.contains(configMatcher) {
-                existingEvent.append(configEntry)
-            }
-        }
-
-        return existingEvent
-    }
-
-    /// Deduplicate managed entries by matcher, merging user hooks from duplicates
-    /// Returns updated entries and set of seen matchers
-    private static func deduplicateClaudeAtollEntries(
-        in entries: [[String: Any]],
-        command: String,
-        eventName: String,
-    ) -> ([[String: Any]], Set<String>) {
-        var result = entries
-        var matcherToFirstIndex: [String: Int] = [:]
-        var indicesToRemove = [Int]()
-
-        for i in result.indices {
-            guard var entryHooks = result[i]["hooks"] as? [[String: Any]],
-                  self.isClaudeAtollHookEntry(entryHooks)
-            else { continue }
-
-            let matcherKey = (result[i]["matcher"] as? String) ?? ""
-
-            if let firstIndex = matcherToFirstIndex[matcherKey] {
-                // Duplicate - merge user hooks into first entry, then mark for removal
-                self.mergeUserHooks(from: entryHooks, into: &result, at: firstIndex, eventName: eventName)
-                indicesToRemove.append(i)
-            } else {
-                // First occurrence - update command and track matcher
-                matcherToFirstIndex[matcherKey] = i
-                self.updateClaudeAtollCommand(in: &entryHooks, to: command)
-                result[i]["hooks"] = entryHooks
-            }
-        }
-
-        // Remove duplicates in reverse order to preserve indices
-        if !indicesToRemove.isEmpty {
-            Self.logger.info("Removed \(indicesToRemove.count) duplicate Claude Atoll hook entry(ies) from \(eventName)")
-            for index in indicesToRemove.reversed() {
-                result.remove(at: index)
-            }
-        }
-
-        return (result, Set(matcherToFirstIndex.keys))
-    }
-
-    /// Remove managed entries from hook events we no longer register on.
-    /// Preserves unrelated entries (e.g. rtk's PreToolUse hooks).
-    /// TODO(anthropics/claude-code#15897): Remove this method once PreToolUse is re-registered.
-    private static func removeDeprecatedHookEntries(from hooks: inout [String: Any]) {
-        let activeEvents = Set(self.buildHookConfigurations(command: "").map(\.0))
-        let deprecatedEvents = ["PreToolUse"]
-
-        for event in deprecatedEvents where !activeEvents.contains(event) {
-            guard var entries = hooks[event] as? [[String: Any]] else { continue }
-
-            // Remove legacy direct format entries
-            entries.removeAll { self.isLegacyDirectEntry($0) }
-
-            // For modern wrapped format: remove managed hooks from each entry,
-            // but preserve entries that have unrelated hooks
-            var indicesToRemove = [Int]()
-            for i in entries.indices {
-                guard var entryHooks = entries[i]["hooks"] as? [[String: Any]] else { continue }
-                let hadClaudeAtoll = entryHooks.contains { hook in
-                    self.isManagedHookCommand(hook["command"] as? String)
-                }
-                guard hadClaudeAtoll else { continue }
-
-                entryHooks.removeAll { hook in
-                    self.isManagedHookCommand(hook["command"] as? String)
-                }
-
-                if entryHooks.isEmpty {
-                    indicesToRemove.append(i)
-                } else {
-                    entries[i]["hooks"] = entryHooks
-                }
-            }
-
-            for index in indicesToRemove.reversed() {
-                entries.remove(at: index)
-            }
-
-            if entries.isEmpty {
-                hooks.removeValue(forKey: event)
-                Self.logger.info("Removed deprecated Claude Atoll hook entries from \(event)")
-            } else {
-                hooks[event] = entries
-                Self.logger.info("Cleaned Claude Atoll hooks from \(event), preserved \(entries.count) other entry(ies)")
-            }
-        }
-    }
-
-    /// Check if hooks array contains a managed hook
-    private static func isClaudeAtollHookEntry(_ hooks: [[String: Any]]) -> Bool {
-        hooks.contains { hook in
-            self.isManagedHookCommand(hook["command"] as? String)
-        }
-    }
-
-    /// Merge unrelated hooks from source into the target entry
-    private static func mergeUserHooks(
-        from sourceHooks: [[String: Any]],
-        into entries: inout [[String: Any]],
-        at targetIndex: Int,
-        eventName: String,
-    ) {
-        let userHooks = sourceHooks.filter { hook in
-            guard let cmd = hook["command"] as? String else { return true }
-            return !self.isManagedHookCommand(cmd)
-        }
-
-        guard !userHooks.isEmpty,
-              var targetHooks = entries[targetIndex]["hooks"] as? [[String: Any]]
-        else { return }
-
-        targetHooks.append(contentsOf: userHooks)
-        entries[targetIndex]["hooks"] = targetHooks
-        Self.logger.info("Merged \(userHooks.count) user hook(s) from duplicate entry in \(eventName)")
-    }
-
-    /// Update managed command in hooks array
-    private static func updateClaudeAtollCommand(in hooks: inout [[String: Any]], to command: String) {
-        for j in hooks.indices where self.isManagedHookCommand(hooks[j]["command"] as? String) {
-            hooks[j]["command"] = command
-        }
-    }
-
-    /// Check if entry is a legacy direct format (type: command at top level, not wrapped in hooks)
-    private static func isLegacyDirectEntry(_ entry: [String: Any]) -> Bool {
-        // Legacy format: {"type": "command", "command": "...claude-atoll-state.py..."}
-        // Modern format: {"hooks": [{"type": "command", "command": "..."}]}
-        if entry["hooks"] != nil {
-            return false // This is the modern wrapped format
-        }
-        if let type = entry["type"] as? String, type == "command",
-           let cmd = entry["command"] as? String,
-           self.isManagedHookCommand(cmd) {
-            return true
-        }
-        return false
-    }
-
-    /// Check if entry contains a Claude Atoll command (either wrapped or direct format)
-    private static func containsClaudeAtollCommand(_ entry: [String: Any]) -> Bool {
-        // Check modern wrapped format: {"hooks": [{"type": "command", "command": "..."}]}
-        if let entryHooks = entry["hooks"] as? [[String: Any]] {
-            for hook in entryHooks {
-                if let cmd = hook["command"] as? String,
-                   self.isManagedHookCommand(cmd) {
-                    return true
-                }
-            }
-        }
-        // Check legacy direct format: {"type": "command", "command": "..."}
-        if self.isLegacyDirectEntry(entry) {
-            return true
-        }
-        return false
-    }
-
-    private static func isManagedHookCommand(_ command: String?) -> Bool {
-        guard let command else { return false }
-        return Self.managedHookScriptNames.contains { command.contains($0) }
-    }
 }
-
-// swiftlint:enable type_body_length
 
 // MARK: - SettingsIO
